removed many states from felleskomponent-selection components and small refactoring caused by this
